requirements.txt
asgiref==3.2.7
autopep8==1.4.4
Babel==2.8.0
bootstrap-admin==0.4.4
certifi==2019.3.9
chardet==3.0.4
chmod==2019.4.13
click==7.1.2
csv342==1.0.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.10
django-allauth==0.39.1
django-bootstrap3==12.1.0
django-ckeditor==5.9.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-extra-views==0.11.0
django-haystack==2.8.1
django-js-asset==1.2.2
django-json-field==0.5.7
django-jsonfield==1.4.0
django-oscar==2.0.4
django-phonenumber-field==2.0.1
django-six==1.0.4
django-tables2==1.21.2
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
factory-boy==2.12.0
Faker==4.1.0
gunicorn==20.0.4
idna==2.8
mysql-connector-python==8.0.20
oauthlib==3.0.1
pep8==1.7.1
phonenumbers==8.12.5
Pillow==7.1.2
protobuf==3.12.2
psutil==5.7.0
public==2019.4.13
purl==1.5
pycodestyle==2.5.0
PyMySQL==0.9.3
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
runcmd==2019.4.13
six==1.15.0
sqlparse==0.2.4
stripe==2.27.0
svgwrite==1.4
text-unidecode==1.3
Tree==0.2.4
upgrade-requirements==1.7.0
urllib3==1.24.2
whitenoise==5.1.0
